refactor: 불필요 코드 제거 (#39)
* Feat/project API 명세 추가 (#13)

* setting: 서비스 예외처리 기본 설정

* feat:project 생성 기능 추가

* feat:project 기본 CRUD 구현

* feat:project 기본 CRUD 구현

* feat: task - 기본 crud 기능 추가 (#5)

* feat: 업무 기본 CRUD 추가

* feat: 업무와 프로젝트 연관 관계 설정 #2

* refactor: 기존 코드 수정
- 페이지네이션 로직 변경
- service 인터페이스 추가

* refactor: 업무 코드 리팩토링 #2
- entity -> dto 전달 방식으로 변경 및 DTO 파일 분할
- 업무 상태 로직 변경으로 불필요 파일 삭제 (TaskMapper, TaskStatus)
- swagger api 코드 추가
- transaction readonly 속성 추가
- entity - setter 제거

* feat: task swagger api 추가

* feat: project CRUD 추가 (#6)

* feat:project 생성 기능 추가

* feat:project 기본 CRUD 구현

* feat:project 기본 CRUD 구현

---------


* Revert "Feat/project API 명세 추가 (#13)" (#15)

This reverts commit abcf68a171716c1c796019a6f98a3d80306e29f2.

* feat:project 내 프로젝트 검색 기능 구현 #12 (#26)

* setting: 서비스 예외처리 기본 설정

* feat: 업무 기본 CRUD 추가

* settings: Cors 및 시큐리티 필터 설정

* feature: jwt 토큰 및 Authentication 생성, 조회 기능 추가

* feat: 업무와 프로젝트 연관 관계 설정 #2

* feature: MemberMapper 파일 작성 및 회원 관련 CRUD 틀 작성

* feat:project 생성 기능 추가

* refactor: 기존 코드 수정
- 페이지네이션 로직 변경
- service 인터페이스 추가

* feature: 회원가입 로직 작성

* feature: jwt 토큰 작업 및 로그인 기능 작성

* feature: Redis 설정 및 refresh 토큰 발급, 제거 기능 추가

* feature: 이메일 인증 클래스 및 테스트

* feat: 회원정보 CRUD 기능구현

* refactor: 업무 코드 리팩토링 #2
- entity -> dto 전달 방식으로 변경 및 DTO 파일 분할
- 업무 상태 로직 변경으로 불필요 파일 삭제 (TaskMapper, TaskStatus)
- swagger api 코드 추가
- transaction readonly 속성 추가
- entity - setter 제거

* feature: 이메일 인증 로직 추가

* feat:project 기본 CRUD 구현

* feat:project 기본 CRUD 구현

* refactor: 환경설정 세팅 및 소프트 삭제 여부 체크 수정

* feat: task swagger api 추가

* feat: task - 기본 crud 기능 추가 (#5)

* feat: 업무 기본 CRUD 추가

* feat: 업무와 프로젝트 연관 관계 설정 #2

* refactor: 기존 코드 수정
- 페이지네이션 로직 변경
- service 인터페이스 추가

* refactor: 업무 코드 리팩토링 #2
- entity -> dto 전달 방식으로 변경 및 DTO 파일 분할
- 업무 상태 로직 변경으로 불필요 파일 삭제 (TaskMapper, TaskStatus)
- swagger api 코드 추가
- transaction readonly 속성 추가
- entity - setter 제거

* feat: task swagger api 추가

* feat: project CRUD 추가 (#6)

* feat:project 생성 기능 추가

* feat:project 기본 CRUD 구현

* feat:project 기본 CRUD 구현

* feat : 코드 리팩토링

* refactor: Token renew 로직 관련 수정

* refactor : 코드 리팩토링 및 오류 수정

* refactor: Service, Repository Layer 추상화

* feat: task status CRUD 추가

* feat: task와 task status 연관관계 설정, project 생성시 기본 task status 4가지 설정 #2

* feat:project API  명세 추가 (#14)

* feat:project 생성 기능 추가

* feat:project 기본 CRUD 구현

* feat:project 기본 CRUD 구현

* feat:project 유효성 로직 추가  #12

* feat : 스웨거 추가

* refactor: Security matcher 및 Jwt 필터 수정

* style: 스타일 변경 #2
- resource - mapper 폴더 위치 변경
- xml : sql문 대문자 변경
- mybatis mapper -> repository DI 않고 service 에서 mapper 바로 사용하도록 변경

* feat: member_task 관련 업무 담당자 추가/삭제 기능 추가

* feat: 프로젝트 생성시 기본 업무 상태 추가 #7

* setting: jasypt 설정 및 yml 파일 암호화

* feat: 업무 담당자 추가 삭제 기능 추가, 프로젝트 생성시 업무 상태 추가 (#17)

* feat: 업무 기본 CRUD 추가

* feat: 업무와 프로젝트 연관 관계 설정 #2

* refactor: 기존 코드 수정
- 페이지네이션 로직 변경
- service 인터페이스 추가

* refactor: 업무 코드 리팩토링 #2
- entity -> dto 전달 방식으로 변경 및 DTO 파일 분할
- 업무 상태 로직 변경으로 불필요 파일 삭제 (TaskMapper, TaskStatus)
- swagger api 코드 추가
- transaction readonly 속성 추가
- entity - setter 제거

* feat: task swagger api 추가

* feat: task status CRUD 추가

* feat: task와 task status 연관관계 설정, project 생성시 기본 task status 4가지 설정 #2

* style: 스타일 변경 #2
- resource - mapper 폴더 위치 변경
- xml : sql문 대문자 변경
- mybatis mapper -> repository DI 않고 service 에서 mapper 바로 사용하도록 변경

* feat: member_task 관련 업무 담당자 추가/삭제 기능 추가

* feat: 프로젝트 생성시 기본 업무 상태 추가 #7

* fix: task status 컨트롤러 코드 수정

* feat: 파일, 파일정보 CRUD  (#20)

* feat: file 기본 세팅

* update

* style: 업무, 업무상태 식별자 타입 변경 Integer -> Long

* feat:comment 기본 CRUD 구현 #18 (#21)

* feat:project 생성 기능 추가

* feat:project 기본 CRUD 구현

* feat:project 기본 CRUD 구현

* feat:project 유효성 로직 추가  #12

* feat:comment 댓글 기본 CRUD #18

* style:불필요한 파일 제거

* feat:comment 댓글 CRUD 기능 구현 #18

* feat: 프로젝트 초대 기본 기능 구현 (#23)

* feat: 프로젝트 초대 코드 생성, 목록 추가 기능 구현

* refactor: RedisService 동일 로직 메서드 공통화

* Revert "feat: 프로젝트 초대 기본 기능 구현 (#23)" (#24)

This reverts commit afbd02124e062e51a72ab68150c99ff8436140a0.

* feat:project 종료일 순 정렬, 이름 검색 기능 추가 (#25)

* feat:project 생성 기능 추가

* feat:project 기본 CRUD 구현

* feat:project 기본 CRUD 구현

* feat:project 유효성 로직 추가  #12

* feat:comment 댓글 기본 CRUD #18

* style:불필요한 파일 제거

* feat:comment 댓글 CRUD 기능 구현 #18

* feat:comment 프로젝트 end_date 순 정렬, 검색 기능 구현 #12

* feat:comment 프로젝트 end_date 순 정렬, 검색 기능 구현 #12

* fix: 컨트롤러 코드 수정

* feat:project 프로젝트 내 프로젝트 조회 기능 구현  #12

---------


* Revert "feat:project 내 프로젝트 검색 기능 구현 #12 (#26)" (#29)

This reverts commit 0eacfe50dfd393ad863825abb758db5c603915af.

* refactor: 기본 pk 타입 변경, 삭제시 수정일자 업데이트, 업무 목록 조회 jpa->mybatis 변경

* style: 불필요 코드 제거, 엔티티 수정
- 생성일, 수정일 코드 수정
- project - progress 제거
- member_project - status 추가

* style: 불필요 코드 제거, 엔티티 수정
- 생성일, 수정일 코드 수정
- project - progress 제거
- member_project - status 추가

* fix: 프로젝트 페이지네이션 오류 보완

---------

+++ src/main/java/com/douzone/prosync/task_status/service/TaskStatusServiceImpl.java
@@ -10,7 +10,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-//TODO : 권한가진 프로젝트회원인지 검증
 @Service
 @Transactional
 @RequiredArgsConstructor
@@ -18,19 +17,23 @@ public class TaskStatusServiceImpl implements TaskStatusService {
 
     private final TaskStatusMapper taskStatusMapper;
 
-    public Integer createTaskStatus(Long projectId, TaskStatusDto.PostDto requestBody, Long memberId) {
+    public Long createTaskStatus(Long projectId, TaskStatusDto.PostDto requestBody) {
         taskStatusMapper.save(projectId, requestBody);
         return requestBody.getTaskStatusId();
     }
 
-    public void updateTaskStatus(Integer taskStatusId, TaskStatusDto.PatchDto requestBody, Long memberId) {
+    public void updateTaskStatus(Long taskStatusId, TaskStatusDto.PatchDto requestBody, Long memberId) {
         requestBody.setTaskStatusId(taskStatusId);
         verifyExistTaskStatus(taskStatusId);
         taskStatusMapper.update(requestBody);
     }
 
-    public void deleteTaskStatus(Integer taskStatusId, Long memberId) {
+    public void deleteTaskStatus(Long taskStatusId, Long memberId) {
         verifyExistTaskStatus(taskStatusId);
+        // task에 삭제하는 task status가 지정되어 있으면 예외 던짐
+        if (taskStatusMapper.findTaskByTaskStatus(taskStatusId) != 0) {
+            throw new ApplicationException(ErrorCode.TASK_EXISTS);
+        }
         taskStatusMapper.delete(taskStatusId);
     }
 
@@ -38,13 +41,18 @@ public class TaskStatusServiceImpl implements TaskStatusService {
         return taskStatusMapper.findTaskStatusByProject(projectId, isActive);
     }
 
-    public TaskStatusDto.GetResponseDto getTaskStatus(Integer taskStatusId, Long memberId) {
-        return taskStatusMapper.findTaskStatus(taskStatusId).orElseThrow(() -> new ApplicationException(ErrorCode.TASK_STATUS_NOT_FOUND));
+    public TaskStatusDto.GetResponseDto getTaskStatus(Long taskStatusId, Long memberId) {
+        TaskStatusDto.GetResponseDto findTaskStatus = findTaskStatus(taskStatusId);
+        return findTaskStatus;
     }
 
-    private void verifyExistTaskStatus(Integer taskStatusId) {
+    private void verifyExistTaskStatus(Long taskStatusId) {
         if (taskStatusMapper.findExistsTaskStatus(taskStatusId) == 0) {
-                throw new ApplicationException(ErrorCode.TASK_STATUS_NOT_FOUND);
+            throw new ApplicationException(ErrorCode.TASK_STATUS_NOT_FOUND);
         }
+    }
+
+    private TaskStatusDto.GetResponseDto findTaskStatus(Long taskStatusId) {
+        return taskStatusMapper.findTaskStatus(taskStatusId).orElseThrow(() -> new ApplicationException(ErrorCode.TASK_STATUS_NOT_FOUND));
     }
 }
